Update most GitHub actions (#64)
This updates most of the GitHub actions. The action to update `setup-gcloud` is not updated as part of this, because the way the authentication is handled changes in newer versions. We should update that as well at some point, because it's an ancient version.

This should hopefully resolve the issue that the validation pipeline fails in recent runs.