Update cwt
Tests on some PRs that don't touch the api component at all are now failing with an error like this:
```
../../../../.cache/pypoetry/virtualenvs/api-7sD7cF5e-py3.11/lib/python3.11/site-packages/oscrypto/_openssl/_libcrypto_cffi.py:44: in <module>
    raise LibraryNotFoundError('Error detecting the version of libcrypto')
E   oscrypto.errors.LibraryNotFoundError: Error detecting the version of libcrypto
```

At least https://community.snowflake.com/s/article/Python-Connector-fails-to-connect-with-LibraryNotFoundError-Error-detecting-the-version-of-libcrypto states that there's an issue with `oscrypto` 1.3.0 and earlier (there's no newer releases of it than 1.3.0 for the record) has some issue with multi digit patch versions of OpenSSL. Most likely the GitHub actions are now running on VMs with such version of OpenSSL (unverified guess).

The 3.0.0 version of `cwt` ditched `oscrypto`, so this PR updates `cwt` to 3.0.0 version. None of the listed changes at https://github.com/dajiaji/python-cwt/blob/main/CHANGES.rst#version-300 indicate any big breaking changes and all the tests are still passing after updating it.